requirements.txt
nose
dateutils
pexpect
